I dont know, i forgot how long ive been inactive on this. imma just assume i fixed the alignment sizes a bit

+++ vyn_parser/src/parse/parse_loop/parse_for_loop.rs
@@ -45,7 +45,7 @@ impl Parser {
             // let i = <starting_num>;
             Stmt::VarDecl {
                 name: id.clone(),
-                typ: Some("int".to_string()),
+                typ: Some("i128".to_string()),
                 value: Some(Expr::Literal(Value::Int(starting_num))),
             },
             // loop {}
